Mark entities in blueprints as 'non-snapping'
When building blueprints, whenever an entity is placed that would "snap"
a miniloader towards it. As a result, depending on the order of building,
miniloaders may end up in a different direction than in the blueprint.

Mark all entities that would cause miniloaders to snap with a specific tag
in the blueprint so that they do not affect miniloaders.